+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * [ENHANCEMENT] Force `lmutil` to run with `LANG=C`.
 * [ENHANCEMENT] Update vendoring.
 * [ENHANCEMENT] Improve testing.
+* [ENHANCEMENT] expose `-i` expiration dates in seconds with new
+  `lmstat_feature_exp` collector.
 * [BUGFIX] Fix `feature_to_include` logic error.
 
 ## v0.0.1 / 2018-06-12